+++ third-party/abseil.zig
@@ -22,8 +22,11 @@ pub fn build(b: *std.Build, config: Config) Dependency {
 
     switch (config.target.result.os.tag) {
         .windows => mod.linkSystemLibrary("dbghelp", .{}),
-        .macos => Dependency.addFrameworkSearchPaths(mod, config.target),
-        else => {}, 
+        .macos => {
+            Dependency.addFrameworkSearchPaths(mod, config.target);
+            mod.linkFramework("CoreFoundation", .{});
+        },
+        else => {},
     }
 
     mod.addIncludePath(root);